@@ -1,18 +1,14 @@
 plugins {
-    id 'application'
+    id 'java-library'
     id 'com.github.johnrengelman.shadow' version '7.1.1'
 
-    // Used by release.gradle
     id 'maven-publish'
-    id 'signing'
-    id 'io.codearte.nexus-staging' version '0.30.0'
+
 }
 
-apply plugin: 'application'
-apply plugin: 'com.github.johnrengelman.shadow'
 
-group 'nl.martijndwars'
-version '5.1.2'
+group 'com.github.xabgesagtx'
+version '0.1'
 
 repositories {
     mavenLocal()
@@ -20,17 +16,9 @@ repositories {
 }
 
 dependencies {
-    // For CLI
-    implementation group: 'com.beust', name: 'jcommander', version: '1.81'
-
-    // For making HTTP requests
-    implementation group: 'org.apache.httpcomponents', name: 'httpasyncclient', version: '4.1.5'
-
-    // For making async HTTP requests
-    implementation group: 'org.asynchttpclient', name: 'async-http-client', version: '2.12.4'
 
     // For cryptographic operations
-    shadow group: 'org.bouncycastle', name: 'bcprov-jdk18on', version: '1.82'
+    api group: 'org.bouncycastle', name: 'bcprov-jdk18on', version: '1.82'
 
     // For creating and signing JWT
     implementation group: 'org.bitbucket.b_c', name: 'jose4j', version: '0.9.6'
@@ -38,20 +26,17 @@ dependencies {
     // For parsing JSON
     testImplementation group: 'com.google.code.gson', name: 'gson', version: '2.8.9'
 
-    // For making HTTP requests
-    testImplementation group: 'org.apache.httpcomponents', name: 'fluent-hc', version: '4.5.13'
-
     // For testing, obviously
-    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.8.1'
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.14.1'
 
     // For running JUnit tests
-    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.8.1'
+    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.14.1'
 
     // For turning InputStream to String
-    testImplementation group: 'commons-io', name: 'commons-io', version: '2.11.0'
+    testImplementation group: 'commons-io', name: 'commons-io', version: '2.21.0'
 
     // For reading the demo vapid keypair from a pem file
-    testImplementation group: 'org.bouncycastle', name: 'bcpkix-jdk15on', version: '1.70'
+    testImplementation group: 'org.bouncycastle', name: 'bcpkix-jdk18on', version: '1.81'
 
     // For verifying Base64Encoder results in unit tests
     testImplementation group: 'com.google.guava', name: 'guava', version: '33.4.8-jre'
@@ -63,17 +48,8 @@ wrapper {
 
 java {
     toolchain {
-        languageVersion = JavaLanguageVersion.of(17)
+        languageVersion = JavaLanguageVersion.of(21)
     }
-}
-
-
-application {
-    mainClass = 'nl.martijndwars.webpush.cli.Cli'
-}
-
-run {
-    classpath configurations.shadow.files
 }
 
 test {
@@ -103,6 +79,3 @@ artifacts {
     archives sourcesJar
 }
 
-if (hasProperty('release')) {
-    apply from: 'release.gradle'
-}
