@@ -4,22 +4,19 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import org.apache.commons.io.IOUtils;
-import org.apache.http.HttpEntity;
-import org.apache.http.client.fluent.Request;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
 
 /**
  * Java wrapper for interacting with the Web Push Testing Service.
  */
-public class TestingService {
-    private String baseUrl;
+public class TestingService implements AutoCloseable {
+    private final String baseUrl;
+    private final HttpClient httpClient = HttpClient.newHttpClient();
 
     public TestingService(String baseUrl) {
         this.baseUrl = baseUrl;
@@ -30,17 +27,12 @@ public class TestingService {
      *
      * @return
      */
-    public int startTestSuite() throws IOException {
+    public int startTestSuite() throws IOException, InterruptedException {
         String startTestSuite = request(baseUrl + "start-test-suite/");
 
         JsonElement root = JsonParser.parseString(startTestSuite);
 
-        return root
-                .getAsJsonObject()
-                .get("data")
-                .getAsJsonObject()
-                .get("testSuiteId")
-                .getAsInt();
+        return root.getAsJsonObject().get("data").getAsJsonObject().get("testSuiteId").getAsInt();
     }
 
     /**
@@ -51,7 +43,7 @@ public class TestingService {
      * @return
      * @throws IOException
      */
-    public JsonObject getSubscription(int testSuiteId, Configuration configuration) throws IOException {
+    public JsonObject getSubscription(int testSuiteId, Configuration configuration) throws IOException, InterruptedException {
         JsonObject jsonObject = new JsonObject();
         jsonObject.addProperty("testSuiteId", testSuiteId);
         jsonObject.addProperty("browserName", configuration.browser);
@@ -65,9 +57,7 @@ public class TestingService {
             jsonObject.addProperty("vapidPublicKey", configuration.publicKey);
         }
 
-        HttpEntity entity = new StringEntity(jsonObject.toString(), ContentType.APPLICATION_JSON);
-
-        String getSubscription = request(baseUrl + "get-subscription/", entity);
+        String getSubscription = request(baseUrl + "get-subscription/", jsonObject);
 
         return getData(getSubscription);
     }
@@ -80,14 +70,13 @@ public class TestingService {
      * @return
      * @throws IOException
      */
-    public JsonArray getNotificationStatus(int testSuiteId, int testId) throws IOException {
+    public JsonArray getNotificationStatus(int testSuiteId, int testId) throws IOException, InterruptedException {
         JsonObject jsonObject = new JsonObject();
         jsonObject.addProperty("testSuiteId", testSuiteId);
         jsonObject.addProperty("testId", testId);
 
-        HttpEntity entity = new StringEntity(jsonObject.toString(), ContentType.APPLICATION_JSON);
 
-        String notificationStatus = request(baseUrl + "get-notification-status/", entity);
+        String notificationStatus = request(baseUrl + "get-notification-status/", jsonObject);
 
         return getData(notificationStatus).get("messages").getAsJsonArray();
     }
@@ -97,13 +86,11 @@ public class TestingService {
      *
      * @return
      */
-    public boolean endTestSuite(int testSuiteId) throws IOException {
+    public boolean endTestSuite(int testSuiteId) throws IOException, InterruptedException {
         JsonObject jsonObject = new JsonObject();
         jsonObject.addProperty("testSuiteId", testSuiteId);
 
-        HttpEntity entity = new StringEntity(jsonObject.toString(), ContentType.APPLICATION_JSON);
-
-        String endTestSuite = request(baseUrl + "end-test-suite/", entity);
+        String endTestSuite = request(baseUrl + "end-test-suite/", jsonObject);
 
         return getData(endTestSuite).get("success").getAsBoolean();
     }
@@ -114,7 +101,7 @@ public class TestingService {
      * @param uri
      * @return
      */
-    protected String request(String uri) throws IOException {
+    protected String request(String uri) throws IOException, InterruptedException {
         return request(uri, null);
     }
 
@@ -124,24 +111,29 @@ public class TestingService {
      * @param uri
      * @return
      */
-    protected String request(String uri, HttpEntity entity) throws IOException {
-        return Request.Post(uri).body(entity).execute().handleResponse(httpResponse -> {
-            String json = EntityUtils.toString(httpResponse.getEntity());
+    protected String request(String uri, JsonObject entity) throws IOException, InterruptedException {
+        HttpRequest.Builder builder = HttpRequest.newBuilder(URI.create(uri));
+        if (entity != null) {
+            builder.POST(HttpRequest.BodyPublishers.ofString(entity.toString())).header("Content-Type", "application/json");
+        } else {
+            builder.POST(HttpRequest.BodyPublishers.noBody());
+        }
+        java.net.http.HttpRequest request = builder.build();
+        HttpResponse<String> httpResponse = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
+        String json = httpResponse.body();
 
-            if (httpResponse.getStatusLine().getStatusCode() != 200) {
-                JsonElement root = JsonParser.parseString(json);
-                JsonObject error = root.getAsJsonObject().get("error").getAsJsonObject();
+        if (httpResponse.statusCode() != 200) {
+            JsonElement root = JsonParser.parseString(json);
+            JsonObject error = root.getAsJsonObject().get("error").getAsJsonObject();
 
-                String errorId = error.get("id").getAsString();
-                String errorMessage = error.get("message").getAsString();
+            String errorId = error.get("id").getAsString();
+            String errorMessage = error.get("message").getAsString();
 
-                String body = IOUtils.toString(entity.getContent(), UTF_8);
 
-                throw new IllegalStateException("Error while requesting " + uri + " with body " + body + " (" + errorId + ": " + errorMessage);
-            }
+            throw new IllegalStateException("Error while requesting " + uri + " with body " + entity + " (" + errorId + ": " + errorMessage);
+        }
 
-            return json;
-        });
+        return json;
     }
 
     /**
@@ -152,9 +144,11 @@ public class TestingService {
     protected JsonObject getData(String response) {
         JsonElement root = JsonParser.parseString(response);
 
-        return root
-                .getAsJsonObject()
-                .get("data")
-                .getAsJsonObject();
+        return root.getAsJsonObject().get("data").getAsJsonObject();
+    }
+
+    @Override
+    public void close() {
+        httpClient.close();
     }
 }
